Use KujakuLocation in TrackingService

+++ library/src/main/java/io/ona/kujaku/location/KujakuLocation.java
@@ -12,7 +12,7 @@ public class KujakuLocation extends Location {
 
     private long tag;
 
-    public KujakuLocation(@NonNull Location location, int tag) {
+    public KujakuLocation(@NonNull Location location, long tag) {
         super(location);
         this.tag = tag;
     }
